finish fibsys and fix "long" in other

+++ mulmod.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
 
 int main(int argc, char ** argv) {
-    long a, b, m;
-    scanf("%ld%ld%ld", &a, &b, &m);
-    long x = b;
+    long long a, b, m;
+    scanf("%lld%lld%lld", &a, &b, &m);
+    long long x = b;
     int s[100];
     int k = 0;
     while(x){
@@ -11,12 +11,12 @@ int main(int argc, char ** argv) {
         x /= 2;
         k++;
     }
-    long res = a * s[k - 1];
+    long long res = a * s[k - 1];
     //printf("%ld ", k);
     for (int i = 0; i <= k - 2; i++){
         res *= 2;
         res = (res % m) + ((a * s[k - i - 2]) % m);
     }
-    printf("%ld", res % m);
+    printf("%lld", res % m);
     return 0;
 }
