+++ polinom.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 
-long long gorner_count(long s[], long x0, long n){
+long long gorner_count(long long s[], long long x0, long long n){
     long long res = s[0];
     for (int i = 0; i < n; i++){
         res *= x0;
@@ -10,19 +10,19 @@ long long gorner_count(long s[], long x0, long n){
 }
 
 int main(int argc, char ** argv) {
-    long x0, s[10000], n;
-    scanf("%ld%ld", &n, &x0);
+    long long x0, s[10000], n;
+    scanf("%lld%lld", &n, &x0);
     for (int i = 0; i < n + 1; i++){
-        scanf("%ld", &s[i]);
+        scanf("%lld", &s[i]);
     }
-    long res1 = gorner_count(s, x0, n);
-    printf("%ld\n", res1);
+    long long res1 = gorner_count(s, x0, n);
+    printf("%lld\n", res1);
 
-    long s1[10000];
+    long long s1[10000];
     for(int i = 0; i < n + 1; i++){
         s1[i] = s[i] * (n - i);
     }
     long long res2 = gorner_count(s1, x0, n - 1);
-    printf("%ld\n", res2);
+    printf("%lld\n", res2);
     return 0;
 }
